+++ sei-ibc-go/modules/apps/transfer/keeper/relay.go
@@ -158,7 +158,7 @@ func (k Keeper) SendTransfer(
 		timeoutTimestamp,
 	)
 
-	if err := k.channelKeeper.SendPacket(ctx, channelCap, packet); err != nil {
+	if err := k.ics4Wrapper.SendPacket(ctx, channelCap, packet); err != nil {
 		return err
 	}