+++ sei-ibc-go/testing/mock/ibc_app.go
@@ -8,6 +8,7 @@ import (
 	"github.com/cosmos/ibc-go/v3/modules/core/exported"
 )
 
+// MockIBCApp contains IBC application module callbacks as defined in 05-port.
 type MockIBCApp struct {
 	OnChanOpenInit func(
 		ctx sdk.Context,
@@ -28,9 +29,8 @@ type MockIBCApp struct {
 		channelID string,
 		channelCap *capabilitytypes.Capability,
 		counterparty channeltypes.Counterparty,
-		version,
 		counterpartyVersion string,
-	) error
+	) (version string, err error)
 
 	OnChanOpenAck func(
 		ctx sdk.Context,
@@ -80,16 +80,4 @@ type MockIBCApp struct {
 		packet channeltypes.Packet,
 		relayer sdk.AccAddress,
 	) error
-
-	// NegotiateAppVersion performs application version negotiation given the provided channel ordering, connectionID, portID, counterparty and proposed version.
-	// An error is returned if version negotiation cannot be performed. For example, an application module implementing this interface
-	// may decide to return an error in the event of the proposed version being incompatible with it's own
-	NegotiateAppVersion func(
-		ctx sdk.Context,
-		order channeltypes.Order,
-		connectionID string,
-		portID string,
-		counterparty channeltypes.Counterparty,
-		proposedVersion string,
-	) (version string, err error)
 }